+++ version/src/client/common121/dev/zymekoh/herzium/mixin/MinecraftMixin.java
@@ -2,6 +2,7 @@ package dev.zymekoh.herzium.mixin;
 
 import com.mojang.blaze3d.platform.InputConstants;
 import dev.zymekoh.herzium.compat.KoHsiumIntegration;
+import dev.zymekoh.herzium.config.HerziumConfig;
 import dev.zymekoh.herzium.gui.HerziumWarningScreen;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.Options;
@@ -34,6 +35,11 @@ abstract class MinecraftMixin {
             method = "onGameLoadFinished",
             at = @At(value = "INVOKE", target = "Ljava/lang/Runnable;run()V"))
     private void herzium$showPerformanceWarning(Runnable showInitialScreen) {
+        if (HerziumConfig.get().startupWarningAcknowledged()) {
+            showInitialScreen.run();
+            return;
+        }
+
         Minecraft minecraft = (Minecraft) (Object) this;
         minecraft.setScreen(new HerziumWarningScreen(showInitialScreen));
     }